The beginning of an overhaul to un-jank the project, too many "temporary fixes" for my taste
